+++ tests/localization_tests/test_fr.py
@@ -59,7 +59,7 @@ class FrTest(AbstractLocalizationTestCase, AbstractTestCase):
 
         d = Pendulum.now().add(seconds=1)
         d2 = Pendulum.now()
-        self.assertEqual('1 seconde après', d.diff_for_humans(d2))
+        self.assertEqual(u'1 seconde après', d.diff_for_humans(d2))
         self.assertEqual('1 seconde avant', d2.diff_for_humans(d))
 
         self.assertEqual('1 seconde', d.diff_for_humans(d2, True))
@@ -79,14 +79,14 @@ class FrTest(AbstractLocalizationTestCase, AbstractTestCase):
         d = Pendulum(2016, 8, 28, 7, 3, 6, 123456)
         self.assertEqual('dimanche', d.format('dddd', formatter='alternative'))
         self.assertEqual('dim', d.format('ddd', formatter='alternative'))
-        self.assertEqual('août', d.format('MMMM', formatter='alternative'))
-        self.assertEqual('août', d.format('MMM', formatter='alternative'))
+        self.assertEqual(u'août', d.format('MMMM', formatter='alternative'))
+        self.assertEqual(u'août', d.format('MMM', formatter='alternative'))
         self.assertEqual('AM', d.format('A', formatter='alternative'))
         self.assertEqual('28e', d.format('Do', formatter='alternative'))
 
         self.assertEqual('07:03', d.format('LT', formatter='alternative'))
         self.assertEqual('07:03:06', d.format('LTS', formatter='alternative'))
         self.assertEqual('28/08/2016', d.format('L', formatter='alternative'))
-        self.assertEqual('28 août 2016', d.format('LL', formatter='alternative'))
-        self.assertEqual('28 août 2016 07:03', d.format('LLL', formatter='alternative'))
-        self.assertEqual('dimanche 28 août 2016 07:03', d.format('LLLL', formatter='alternative'))
+        self.assertEqual(u'28 août 2016', d.format('LL', formatter='alternative'))
+        self.assertEqual(u'28 août 2016 07:03', d.format('LLL', formatter='alternative'))
+        self.assertEqual(u'dimanche 28 août 2016 07:03', d.format('LLLL', formatter='alternative'))